+++ docs/app/components/content/FieldGroup.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="mt-5 space-y-5 divide-y divide-(--ui-border) [&>div]:pt-5 [&>div:first-child]:pt-0 [&>div:first-child]:mt-0">
-    <slot />
-  </div>
-</template>