+++ src/runtime/blob/server/utils/blob.ts
@@ -508,6 +508,11 @@ function createMultipartUploadHandler(
       pathname: z.string().min(1)
     }).parse)
 
+    options ||= {}
+    if (getHeader(event, 'x-nuxthub-file-content-type')) {
+      options.contentType ||= getHeader(event, 'x-nuxthub-file-content-type')
+    }
+
     try {
       const object = await createMultipartUpload(pathname, options)
       return {
@@ -645,7 +650,7 @@ function getContentType(pathOrExtension?: string) {
 function mapR2ObjectToBlob(object: R2Object): BlobObject {
   return {
     pathname: object.key,
-    contentType: object.httpMetadata?.contentType,
+    contentType: object.httpMetadata?.contentType || getContentType(object.key),
     size: object.size,
     httpEtag: object.httpEtag,
     uploadedAt: object.uploaded,